requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.4
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
flake8==3.9.2
mccabe==0.6.1
phonenumbers==8.12.26
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
